Refactoring caching tests and adding awaitility to those tests
