+++ modules/ballerina-test/src/test/resources/testng.xml
@@ -41,6 +41,7 @@ under the License.
             <package name="org.ballerinalang.test.parser.*"/>
             <package name="org.ballerinalang.test.launch.*"/>
             <package name="org.ballerinalang.test.task.*"/>
+            <package name="org.ballerinalang.test.caching.*"/>
         </packages>
     </test>
 </suite>
